Unit Test - Adjustments
