--- TheMovieDb/Networking/Result.swift
@@ -1,13 +1,0 @@
-//
-//  Result.swift
-//  TheMovieDb
-//
-//  Created by Misael Chávez on 31/10/21.
-//
-
-import Foundation
-
-enum Result<T, U> where U: Error {
-    case success(T)
-    case failure(U)
-}
